[WIP][profiler] initialize kineto before doing any profiling.
As suggested in robieta in #75504, this workaround seems to fix the cudagraph profiling issues. It moves the kineto initialization step to torch initialization, so that it gets called before any cuda graphs are created.

Tested locally and verified that the issue goes away (although - locally, the symptom is a hanging process, not an illegal memory access).

ghstack-source-id: 6602b993d8ce3f77201ceb764b7c5fed818aba07
Pull Request resolved: https://github.com/pytorch/pytorch/pull/97199