[inductor] Add an AOT mode for the Triton backend
Summary: This PR implements a two-pass wrapper codegen for the Triton
backend to achieve ahead-of-time compilation. In the first pass, the
regular python wrapper code will be generated, and then the generated
code will be executed to perform Triton compilation and autotuning.
After that, the second pass wrapper codegen will generate C++ wrapper
with proper CUDA API to load and launch Triton-generated CUDA kernels.

Like the AOT mode for the cpp backend, the next step would be to provide
a more complete API for AOT.

ghstack-source-id: 0f3774ff5a5127cbe140fcd9ec00361bfd54c1e0
Pull Request resolved: https://github.com/pytorch/pytorch/pull/97152

	modified:   torch/_inductor/codegen/wrapper.py